Rewrite value/raw_value setting
This killed setting .value/.raw_value directly, .set_(raw_)value(),
.state, .serialize(), inheriting_property() and unimethod().

This introduced the .set() method on elements.

Every value is now considered to be "raw" but only unserialized if it is
not "native" to the element, this eliminates .state and .serialize().
.value and .raw_value should now not be set directly anymore and for
COntainers they indeed cannot be set directly anymore, killing
inheriting_property and thereby unimethod()

This change allowed removing a lot of rather embarrasingly complicated
fragile hard to debug code. Making it significantly easier to introduce
new element types and makes it practical for users to create their own
element types that are not just simple derivatives of existing ones.